% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lincs.R
\name{lincs_manifest}
\alias{lincs_manifest}
\title{LINCS reference manifest}
\usage{
lincs_manifest()
}
\value{
A data frame containing the frozen LINCS Level 5 files used by
  ProReT.
}
\description{
LINCS reference manifest
}
